Move workspace destruction and reorder to RESTful resource controllers
