+++ saas/app/controllers/saas/workspace_settings_controller.rb
@@ -3,7 +3,6 @@
 module Saas
   class WorkspaceSettingsController < ApplicationController
     before_action :set_workspace
-    before_action :ensure_administrator, only: [ :destroy ]
 
     def show
       @membership = Current.workspace_membership
@@ -11,25 +10,10 @@ module Saas
       @member_count = User.active.count
     end
 
-    def destroy
-      unless params[:confirmation] == @workspace.name
-        redirect_to settings_path, alert: "Workspace name did not match."
-        return
-      end
-
-      name = @workspace.name
-      @workspace.destroy_with_database!
-      redirect_to workspaces_url(script_name: ""), notice: "#{name} has been deleted"
-    end
-
     private
 
       def set_workspace
         @workspace = Current.workspace
-      end
-
-      def ensure_administrator
-        head :forbidden unless Current.user.administrator?
       end
   end
 end
