+++ saas/test/controllers/saas/workspace_settings_controller_test.rb
@@ -68,63 +68,5 @@ module Saas
 
       assert_response :not_found
     end
-
-    test "destroy requires administrator role" do
-      # Sign in as a non-admin member using create_user! for proper setup
-      charlie_membership = WorkspaceMembership.create!(
-        global_identity: global_identities(:charlie),
-        tenant: @workspace.external_id.to_s
-      )
-      charlie_membership.create_user!(role: :member)
-
-      sign_in_global_identity(global_identities(:charlie))
-
-      workspace_delete "/settings", workspace: @workspace, params: { confirmation: @workspace.name }
-
-      assert_response :forbidden
-    end
-
-    test "destroy requires matching confirmation" do
-      workspace_delete "/settings", workspace: @workspace, params: { confirmation: "wrong name" }
-
-      assert_redirected_to "/#{@workspace.external_id}/settings"
-      assert_equal "Workspace name did not match.", flash[:alert]
-      assert Workspace.exists?(id: @workspace.id)
-    end
-
-    test "destroy with correct confirmation deletes workspace" do
-      external_id = @workspace.external_id
-
-      workspace_delete "/settings", workspace: @workspace, params: { confirmation: @workspace.name }
-
-      assert_redirected_to workspaces_url(script_name: "")
-      assert_match /deleted/, flash[:notice]
-      assert_not Workspace.exists?(external_id: external_id)
-
-      @workspace = nil  # Prevent teardown from trying to destroy
-    end
-
-    test "destroy emails all admins" do
-      # Add a second admin
-      bob_membership = WorkspaceMembership.create!(
-        global_identity: global_identities(:bob),
-        tenant: @workspace.external_id.to_s
-      )
-      bob_membership.create_user!(role: :administrator)
-
-      perform_enqueued_jobs do
-        workspace_delete "/settings", workspace: @workspace, params: { confirmation: @workspace.name }
-      end
-
-      alice_email = global_identities(:alice).email_address
-      bob_email = global_identities(:bob).email_address
-      deletion_emails = ActionMailer::Base.deliveries.select { |m| m.subject.include?("deleted") }
-
-      assert_equal 2, deletion_emails.size
-      assert_equal [ alice_email ], deletion_emails.find { |m| m.to.include?(alice_email) }.to
-      assert_equal [ bob_email ], deletion_emails.find { |m| m.to.include?(bob_email) }.to
-
-      @workspace = nil
-    end
   end
 end
